vite: use a richer diagram, add overlays + rename file into index.ts (usual entry point)
