+++ projects/typescript-vanilla-with-vitejs/README.md
@@ -12,7 +12,7 @@ You will see the following diagram:
 
 ![BPMN diagram in the home page](docs/home.png)
 
-The code calling `bpmn-visualization` to render the BPMN diagram is available in [main.ts](src/main.ts).
+The code calling `bpmn-visualization` to render the BPMN diagram is available in [index.ts](src/index.ts).
 
 If you want to bundle the application, run `npm run build` and then run `npm run preview` to access to a preview of the
 bundle application with http://localhost:4173.
