+++ projects/typescript-vanilla-with-vitejs/index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>bpmn-visualization TypeScript Integration with Vite</title>
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="/src/index.ts"></script>
   </head>
   <body>
     <h1 class="title">bpmn-visualization TypeScript Integration with Vite</h1>
